Move check_veto()
