+++ test/test_readcalls.c
@@ -108,36 +108,6 @@ void test_lookup_in_pfxnummult(void **state) {
 }
 
 
-/* test check_veto() */
-void test_veto_eclude_none(void **state) {
-    assert_int_equal(check_veto(), false);
-}
-
-void test_veto_exclude_country(void **state) {
-    exclude_multilist_type = EXCLUDE_COUNTRY;
-    countrynr = getctynr("HB9ABC");
-    assert_int_equal(check_veto(), false);
-    countrynr = getctynr("DL1AAA");
-    assert_int_equal(check_veto(), true);
-}
-
-void test_veto_exclude_continent_contlist_only(void **state) {
-    continentlist_only = true;
-    exclude_multilist_type = EXCLUDE_CONTINENT;
-    strcpy(continent, "EU");
-    assert_int_equal(check_veto(), false);
-    strcpy(continent, "AF");
-    assert_int_equal(check_veto(), false);
-}
-
-void test_veto_exclude_continent(void **state) {
-    exclude_multilist_type = EXCLUDE_CONTINENT;
-    strcpy(continent, "EU");
-    assert_int_equal(check_veto(), true);
-    strcpy(continent, "AF");
-    assert_int_equal(check_veto(), false);
-}
-
 /* test readcalls */
 void test_add_to_worked(void **state) {
     write_log(LOGFILE);
